+++ druid-lookups/src/test/java/com/yahoo/maha/maha_druid_lookups/server/lookup/namespace/schema/flatbuffer/AdFlatBuffer.fbs
@@ -8,5 +8,6 @@ table ProductAd {
          image_url_large : string;
          source_id : string;
          title : string;
+         last_updated: string;
 }
 root_type ProductAd;
